# docs/swagger.yaml
basePath: /api
definitions:
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.Follower:
    properties:
      follow_id:
        type: string
      followed_at:
        type: string
      id:
        type: string
      user_id:
        type: string
    type: object
  models.Post:
    properties:
      category:
        $ref: '#/definitions/models.PostCategory'
      category_id:
        description: Agora é ObjectID
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      tags:
        description: Lista de Tags
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - category_id
    type: object
  models.PostCategory:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  models.PostReaction:
    properties:
      created_at:
        type: string
      id:
        type: string
      post_id:
        type: string
      reaction:
        description: '"like" ou "dislike"'
        type: boolean
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: seuemail@example.com
    name: Seu Nome
    url: http://www.seusite.com
  description: API para gerenciar posts, comentários, categorias e usuários.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Blog API
  version: "1.0"
paths:
...
  /followers:
    post:
      consumes:
      - application/json
      description: Seguir um usuário
      parameters:
      - description: Seguidor a ser adicionado
        in: body
        name: follower
        required: true
        schema:
          $ref: '#/definitions/models.Follower'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Follower'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Seguir um usuário
      tags:
      - Followers
  /followers/{user_id}:
    get:
      consumes:
      - application/json
      description: Listar seguidores de um usuário
      parameters:
      - description: ID do usuário
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Follower'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Listar seguidores de um usuário
      tags:
      - Followers
...
  /posts:
    get:
      description: Retorna uma lista de posts do banco de dados
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Listar posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Cria um novo post e salva no banco de dados
      parameters:
      - description: Dados do post
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/models.Post'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Post'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Criar um post
      tags:
      - posts
  /posts/{id}:
    delete:
      description: Deleta um post específico do banco de dados
      parameters:
      - description: ID do post
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Deletar post por ID
      tags:
      - posts
    get:
      description: Retorna um post específico do banco de dados
      parameters:
      - description: ID do post
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Buscar post por ID
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Atualiza um post existente no banco de dados
      parameters:
      - description: ID do post
        in: path
        name: id
        required: true
        type: string
      - description: Dados do post
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/models.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Atualizar um post
      tags:
      - posts
  /posts/category/{category_id}:
    get:
      description: Retorna uma lista de posts de uma categoria específica
      parameters:
      - description: ID da categoria
        in: path
        name: category_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Buscar posts por categoria
      tags:
      - posts
...
  /reactions/{id}:
    get:
      consumes:
      - application/json
      description: Buscar reação por ID
      parameters:
      - description: ID da reação
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PostReaction'
      summary: Buscar reação por ID
      tags:
      - Reactions
  /reactions/dislikes/{post_id}:
    get:
      consumes:
      - application/json
      description: Contar total de dislikes de um post
      parameters:
      - description: ID do post
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Total de dislikes
          schema:
            type: string
      summary: Contar total de dislikes de um post
      tags:
      - Reactions
...
  /reactions/likes/{post_id}:
    get:
      consumes:
      - application/json
      description: Contar total de likes de um post
      parameters:
      - description: ID do post
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Total de likes
          schema:
            type: string
      summary: Contar total de likes de um post
      tags:
      - Reactions
...
swagger: "2.0"